Update Contributions.md
originally, the main section of contributions listed the Hugo library version the contribution was written for (since some of the IF Archive ones go many years back), but as I added more contributions to it, I started accidentally putting in the version of the contribution itself, confusing things.  I have now deleted all references to versions in the "stable" section but am keeping them in the "in progress" section.